Begin generalising conventions
Adds HMRC Overseas Trade example and starts removing hard-coded config
specific to the regional-trade example as per #22.

The `is-dimension?` and `is-attribute?` sets are now drawn from the
conventions - i.e. those columns with the respective component
attachment property. The `values` seq now uses an equivalent `is-value?`
set which includes those columns for which the conventions don't specify
a component attachment (i.e. if the column is not a dimension, measure,
or attribute then it must be a value).

`standardise-measure` is now just `title->name`

`slugize-columns` is replaced by `transform-columns` which is configured
by a new convention: `value_transformation` which also allows us to
specify e.g. a `replace-symbols` transform (#18 may later allow us to
derive these instead of setting them explicitly).