Avoid writing tabular files within pipelines.
Add implementations of the csv2rdf RowSource protocol which allow
transformed versions of pipeline input files to be passed directly.

The RowSource protocol represents tabular resources as a logical
sequence of records, each containing the source row number and parsed
data cells. Pipelines previously wrote transformed version of the
input files to disk so they could be passed to the CSVW process.

Add implementations of hte RowSource protocol which allow the
transformation process to be done in memory and presents the
transformed row records directly into the CSVW process.

The number of component specifications derived within cube-pipeline
is expected to be quite small. Load these into memory and add
a RowSource implementation which returns the corresponding tabular
rows to csv2rdf.

Update the tests which check the format of the intermediate
transformed data to used the transformed row sources.