Generate columns in intermediate codelist CSV for top concepts.
Add columns to the intermediate codelist CSV which indicates whether
the current row is a top concept of the schema (i.e. its Parent Notation
is column is empty). Update the codelist metadata to output
skos:hasTopConcept and skos:topConceptOf statements for the corresponding
columns when they are non-empty. Remove the virtual columns that were
previously used to unconditionally output the skos:hasTopConcept and
skos:topConceptOf statements.