Add base-uri parameters to pipelines.
Issue #45 - Add base-uri parameters to each pipeline so it can be
specified in the same way as other parameter in the CLI.

Remove the unused column-config parameters from the components and
codelists pipelines.

Remove the domain key from the configuration and rename
load-configuration to load-column-configuration to indicate the more
specialised usage. Remove the no-arg overload which loads the
configuration from columns.csv.