Convert tabular files into URIs when constructing metadata.
Issue #77 - Use the URI representation of the corresponding tabular
data file when building csv2rdf metadata. Pipeline metadata functions
(.e.g codelist/codelist-metadata) expect to be passed the URI of the
corresponding tabular file. The file object for the temporary
intermediate file was previously being passed directly. On *nix
systems this file is an absolute path to the file
e.g. /var/tmp/file.csv. When this path is resolved against the metadata
URI, this still correctly refers to the temp file path. Absolute
Windows paths are not valid URIs due to the leading directory segment,
so these fail to resolve, resulting in the tabular file not being
located.