Add configuration.column namespace.
Add configuration.column namespace which describes the structure of
a column definition loaded from a columns configuration source.
Add a :type key which indicates whether the column represents a
dimension,measure,attribute or value column. Remove the
component_attachment key from which the type is derived. Add basic
specs for the keys in the column definition and make nillable
columns optional.