Extract URI template definitions.
Issue #69 - Define meta-templates for URIs used in each pipeline.
Meta-templates are URI templates with variable placeholders of the
form $(varname) which are expanded with required pipeline parameters
to create the final templates used within the CSVW output.

The URI templates for each pipeline is defined within their own
resource files.