Validate input CSV files when reading.
Issue #37 - Create a CSV parser which validates and transforms data
rows according to a declarative specification. Use this parser when
parsing input files for the codelist and components pipelines as well
as the columns configuration file within the cube pipeline. This
parser checks that all required columns are present in the input and
no unknown columns are specified.

Rename csv/read-csv-records to read-csv-maps and name the new parser
read-csv-records.