Adopt <type>-records and <type>-schema function-naming convention
This refactoring renames these functions for consistency.

Sequences of record-maps follow the convention "<type>-records":

observation-records (unchanged)
read-component-specifications -> component-specification-records
codes -> code-records
components -> component-records

The maps describing the metadata tableSchema follow the convention "<type>-schema":

used-codes-codes-table -> used-codes-codes-schema
used-codes-codelists-table -> used-codes-codelists-schema
observations-table -> observations-schema
data-structure-definition-table -> data-structure-definition-schema
component-specification-metadata-table -> component-specification-schema
dataset-metadata-table -> dataset-schema
codelist-metadata -> codelist-schema
components-metdata -> components-schema

The term "table" was confusing. A csvw:table is a key in the metadata specification for the table schema but, since we also deal with the tables of records themselves this was confusing - is "table" the data or the metadata about it?!

The terms "metadata" is also overloaded. It could mean the csv schema or the notion that e.g. a qb:DataSet has metadata.