Introduce notation for custom component uris
In order to fully control component URIs we need to introduce another field to add custom slugs. We can otherwise only choose variables from the input or those derived by table2qb.

This patch allows us to specify a `Notation` column in the input to the components pipeline (the name was chosen to match the similar column in the codelist-pipeline). This can then be used to provide any arbitrary slug for the URI. If it's missing we default to slugging the label, which was the pre-existing behaviour.

This allows us to create components like this where the URI slug differs from the label:

    <https://id.milieuinfo.be/def#tijdsperiode> a qb:DimensionProperty; rdfs:label "periode in de tijd" .