Add example of customising dataset/ dsd uris
Note that the actual URI for the DSD is:

    https://id.milieuinfo.be/imjv/dsd/luchtemissie

And not the value in the test:

    https://id.milieuinfo.be/imjv/dsd/luchtemissies

We've used the pluralised version here as it allows us to re-use the same slug that's used in the dataset URI.

We could've reproduced the actual URI by specifying the whole of each URI in the template file, but this would mean the template configuration would only work for this one dataset, instead of allowing this to be reused and tailored to each cube with the `dataset-slug`.

We could also have had the `dataset-slug` not include the plural, and have this appended in the `dataset-uri` only.