Add skos:narrower relations to codelist pipeline
The `parent_notation2` field may seem redundant but it's not. If we
create a virtual column using the same templates then it will
erroneously create triples where the parent_notation is blank. By
repeating the 'real' column we can use it to control whether a triple
should be created (much like the "has_top_concept" and "top_concept_of"
control columns).